+++ Calculadora/Operacoes.cs
@@ -3,10 +3,10 @@ namespace Calculadora
 {
     public class Operacoes
     {
-        public int valorA {get;set;}
+        public long valorA {get;set;}
         public char operador { get; set; }
-        public int valorB { get; set; }
-        public decimal resultado { get; set; }
+        public long valorB { get; set; }
+        public long resultado { get; set; }
 
     }
 }
